# Optimize label sorting