Previously we were sorting labels in every container to build a map key, but also when generating the hash key by calling the `LabelsToSignature` method.  This moves the sorting to be done early in the event processing then passes the sorted label array around.

Before:
```
$ /usr/local/bin/go test -benchmem -run=^$ github.com/prometheus/statsd_exporter -bench '^(BenchmarkHashNameAndLabels)$' -benchtime 5s
goos: darwin
goarch: amd64
pkg: github.com/prometheus/statsd_exporter
BenchmarkHashNameAndLabels/no_labels-8          300000000               26.7 ns/op             0 B/op          0 allocs/op
BenchmarkHashNameAndLabels/one_label-8          50000000               163 ns/op              48 B/op          2 allocs/op
BenchmarkHashNameAndLabels/many_labels-8        10000000               889 ns/op             192 B/op          2 allocs/op
PASS
ok      github.com/prometheus/statsd_exporter   28.928s
```

After:
```
$ /usr/local/bin/go test -benchmem -run=^$ github.com/prometheus/statsd_exporter -bench '^(BenchmarkHashNameAndLabels)$' -benchtime 5s
goos: darwin
goarch: amd64
pkg: github.com/prometheus/statsd_exporter
BenchmarkHashNameAndLabels/no_labels-8          300000000               21.5 ns/op             0 B/op          0 allocs/op
BenchmarkHashNameAndLabels/one_label-8          100000000               59.8 ns/op             0 B/op          0 allocs/op
BenchmarkHashNameAndLabels/many_labels-8        20000000               507 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/prometheus/statsd_exporter   25.530s
```